classifier/requirements.txt
# requirements.txt
kafka-python
sklearn
scikit-learn